ffanalytics/requirements.txt
-f /usr/share/pip-wheels
alembic==1.0.0
asn1crypto==0.24.0
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
Click==7.0
configparser==3.5.0
cryptography==2.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
WTForms==2.2.1
